delete unused prediction file  and rename vector_prediction
