+++ src/mercs/algo/__init__.py
@@ -1,3 +1,1 @@
-from .prediction import (mi, mrai, it, rw)
-
 from .selection import base_selection_algorithm, random_selection_algorithm
